+++ src/ui/action-popup/index.ts
@@ -4,6 +4,7 @@ import { pinia } from "@/utils/pinia"
 import { appRouter } from "@/utils/router"
 import { createApp } from "vue"
 import App from "./app.vue"
+import "@/assets/main.css"
 import "./index.scss"
 
 appRouter.addRoute({